Finish reason for OOP - Write instructions to create new robot "method" - Write instructions to create new robot "attribute" for the method - Write the problem with our approach - Switch the speed to an integer because it makes more sense

+++ app/src/main/java/_training/Robots.java
@@ -1,14 +1,14 @@
 package _training;
 
 public class Robots {
-    static void driveRobot(String direction, double speed){
+    static void driveRobot(String direction, int speed){
         System.out.println("Robot is driving "+direction);
     }
     static void shootRobot(String direction){
         System.out.println("Robot is shooting to the "+direction);
     }
     public static void main(String[] args){
-        double robotSpeed = 20; // Set the robot speed
+        int robotSpeed = 20; // Set the robot speed
         driveRobot("forward", robotSpeed); // Drive the robot forward
         shootRobot("left"); // Shoot the robot to the left
     }
